git env documentation update

+++ src/lib/storage/storage.ts
@@ -3,24 +3,32 @@ import platformConfig from '@/config/app-config';
 import { createClient } from '@supabase/supabase-js'
 
 
-if (!platformConfig.variables.SUPABASE_URL) {
-    throw new Error('SUPABASE_URL is not defined')
-}
-if (!platformConfig.variables.SUPABASE_ANON_KEY) {
-    throw new Error('SUPABASE_ANON_KEY is not defined')
-}
 
 
-// Create a single supabase client for interacting with your database
-const supabase = createClient(platformConfig.variables.SUPABASE_URL, platformConfig.variables.SUPABASE_ANON_KEY)
 
 
 const getClient = () => {
+    // Create a single supabase client for interacting with your database
+    if (!platformConfig.variables.SUPABASE_URL) {
+        console.error('SUPABASE_URL is not defined')
+        return null
+    }
+    if (!platformConfig.variables.SUPABASE_ANON_KEY) {
+        console.error('SUPABASE_ANON_KEY is not defined')
+
+        return null
+    }
+    const supabase = createClient(platformConfig.variables.SUPABASE_URL, platformConfig.variables.SUPABASE_ANON_KEY)
+
     return supabase
 }
 
 const checkBucketExists = async (bucketName: string) => {
 
+    const supabase = getClient()
+    if (!supabase) {
+        return false
+    }
     const { data, error } = await supabase.storage.from(bucketName).list()
     if (error) {
         console.error(error)
@@ -30,6 +38,11 @@ const checkBucketExists = async (bucketName: string) => {
 }
 
 const createBucket = async (bucketName: string) => {
+    const supabase = getClient()
+    if (!supabase) {
+        return false
+    }
+
     const { data, error } = await supabase.storage.createBucket(bucketName)
     if (error) {
         console.error(error)
@@ -39,6 +52,10 @@ const createBucket = async (bucketName: string) => {
 }
 
 const deleteBucket = async (bucketName: string) => {
+    const supabase = getClient()
+    if (!supabase) {
+        return false
+    }
     const { data, error } = await supabase.storage.deleteBucket(bucketName)
     if (error) {
         console.error(error)
@@ -47,6 +64,10 @@ const deleteBucket = async (bucketName: string) => {
     return data
 }
 const uploadFile = async (bucketName: string, path: string, file: File, options: any) => {
+    const supabase = getClient()
+    if (!supabase) {
+        return false
+    }
     const { data, error } = await supabase.storage.from(bucketName).upload(path, file, options)
     if (error) {
         console.error(error)
@@ -56,6 +77,10 @@ const uploadFile = async (bucketName: string, path: string, file: File, options:
 }
 
 const getDownloadUrl = async (bucketName: string, path: string) => {
+    const supabase = getClient()
+    if (!supabase) {
+        return false
+    }
     const { data } = await supabase.storage.from(bucketName).getPublicUrl(path)
     return data
 }
